B4 (#89) cycle 1: relocate WeekFatigueSignals to Models/
Pure refactor. Future cycles add this type to TraineeModelDigest;
the digest lives in Models/ and shouldn't depend on a Services-layer
type. No behavior change. Build succeeds; single definition.
